requirements.txt
cupy_cuda12x==13.3.0
kernel_tuner==1.1.2
lark==1.2.2
matplotlib==3.10.1
numpy==2.2.5
nvidia_ml_py==12.570.86
pandas==2.2.3
pycuda==2025.1
pynvml==12.0.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
statsmodels==0.14.4
torch==2.7.0
